Minor refactoring for Transfer page
